+++ apps/examples/react-vite/src/Footer/Footer.tsx
@@ -2,9 +2,8 @@ import {
   Icon,
   LinkList,
   LinkListLink,
-  Paragraph,
   Footer as RhcFooter,
-} from '@rijkshuisstijl-community/components-react';
+} from '@rijkshuisstijl-community/components-react/no-side-effects';
 import './Footer.css';
 
 export function Footer() {
@@ -17,22 +16,7 @@ export function Footer() {
       columns={[
         {
           appearanceLevel: 3,
-          children: (
-            <>
-              <Paragraph>
-                Dit is geen officiële website van de makers van de Rijkshuisstijl. Dit project wordt niet gesteund door
-                het Ministerie van Algemene Zaken.
-              </Paragraph>
-              <LinkListLink href="#" icon={<Icon icon="chevron-right" />}>
-                De officiele Rijkshuisstijl website
-              </LinkListLink>
-            </>
-          ),
-          heading: 'Disclaimer',
-        },
-
-        {
-          appearanceLevel: 3,
+          heading: 'Over de Rijkshuisstijl Community',
           children: (
             <LinkList>
               <LinkListLink href="#" icon={<Icon icon="chevron-right" />}>
@@ -52,10 +36,10 @@ export function Footer() {
               </LinkListLink>
             </LinkList>
           ),
-          heading: 'Over de Rijkshuisstijl Community',
         },
         {
           appearanceLevel: 3,
+          heading: 'Contact',
           children: (
             <LinkList>
               <LinkListLink href="#" icon={<Icon icon="chevron-right" />}>
@@ -72,7 +56,6 @@ export function Footer() {
               </LinkListLink>
             </LinkList>
           ),
-          heading: 'Contact',
         },
       ]}
     />