+++ packages/components-react/code-input-react/src/index.ts
@@ -1,3 +1,4 @@
 import '@rijkshuisstijl-community/code-input-css';
+import '@rijkshuisstijl-community/text-input-react';
 
 export * from './CodeInput';