+++ packages/components-react/file-input-react/src/index.ts
@@ -1,3 +1,6 @@
+import '@rijkshuisstijl-community/button-react';
 import '@rijkshuisstijl-community/file-input-css';
+import '@rijkshuisstijl-community/file-react';
+import '@rijkshuisstijl-community/paragraph-react';
 
 export * from './FileInput';