+++ packages/components-react/footer-react/src/index.ts
@@ -1,3 +1,7 @@
+import '@rijkshuisstijl-community/column-layout-react';
 import '@rijkshuisstijl-community/footer-css';
+import '@rijkshuisstijl-community/heading-react';
+import '@rijkshuisstijl-community/icon-react';
+import '@rijkshuisstijl-community/link-react';
 
 export * from './Footer';