+++ packages/components-react/library-react/src/FormFieldSelect.tsx
@@ -84,7 +84,7 @@ export const FormFieldSelect = ({
           [statusId]: status,
         }) || undefined
       }
-      className={clsx('rhc-select', {
+      className={clsx('rhc-form-field-select', {
         'utrecht-select--html-select-rtl': dir === 'rtl',
       })}
       onBlur={onBlur}