+++ packages/components-react/pre-heading-react/src/index.ts
@@ -1,0 +1,4 @@
+import '@rijkshuisstijl-community/heading-react';
+import '@rijkshuisstijl-community/pre-heading-css';
+
+export * from './PreHeading';