+++ packages/components-react/radio-group-react/src/index.ts
@@ -1,0 +1,3 @@
+import '@rijkshuisstijl-community/radio-group-css';
+
+export * from './RadioGroup';