+++ packages/components-css/footer-css/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rijkshuisstijl-community/footer-css
 
+## 1.0.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+
 ## 1.0.0
 
 ### Major Changes