+++ packages/components-css/library-css/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rijkshuisstijl-community/components-css
 
+## 17.1.2
+
+### Patch Changes
+
+- Updated dependencies [8d20001]
+  - @rijkshuisstijl-community/pre-heading-css@1.0.0
+
+## 17.1.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+- Updated dependencies [e41c59f]
+  - @rijkshuisstijl-community/language-navigation-css@1.0.1
+  - @rijkshuisstijl-community/listbox-css@1.0.1
+  - @rijkshuisstijl-community/footer-css@1.0.1
+  - @rijkshuisstijl-community/file-css@1.0.1
+
 ## 17.1.0
 
 ### Minor Changes