+++ packages/components-react/code-input-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rijkshuisstijl-community/code-input-react
 
+## 1.0.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+- Updated dependencies [e41c59f]
+  - @rijkshuisstijl-community/text-input-react@1.0.1
+
 ## 1.0.0
 
 ### Major Changes