+++ packages/components-react/language-navigation-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rijkshuisstijl-community/language-navigation-react
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [b3db8e1]
+  - @rijkshuisstijl-community/icon-react@3.1.2
+
+## 1.0.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+- Updated dependencies [e41c59f]
+  - @rijkshuisstijl-community/language-navigation-css@1.0.1
+  - @rijkshuisstijl-community/link-button-react@1.0.1
+  - @rijkshuisstijl-community/icon-react@3.1.1
+  - @rijkshuisstijl-community/link-react@1.1.1
+
 ## 1.0.0
 
 ### Major Changes