+++ proprietary/design-tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rijkshuisstijl-community/design-tokens
 
+## 16.1.0
+
+### Minor Changes
+
+- 9cfac08: - Page Number Navigation tokens van Amsterdam bijgewerkt: juiste types toegekend en missende tokens toegevoegd.
+
+## 16.0.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+
 ## 16.0.0
 
 ### Major Changes